Move SQLExecutionUnit and ConnectionMode to engine package
