+++ shardingsphere-infra/shardingsphere-infra-executor/src/main/java/org/apache/shardingsphere/infra/executor/sql/execute/engine/ConnectionMode.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.executor.sql.execute;
+package org.apache.shardingsphere.infra.executor.sql.execute.engine;
 
 /**
  * Connection Mode.
